点歌/服务端/requirements.txt
fastapi
tqdm
request